fix: unit test

+++ test/integration/04_BaseIntTest_SetOracle.i.sol
@@ -63,7 +63,7 @@ abstract contract BaseIntTest_SetOracle is BaseIntTest_SetMarkets {
 
     // Set AssetPriceConfig
     uint32 _confidenceThresholdE6 = 250; // 2.5% for test only
-    uint256 _trustPriceAge = type(uint256).max; // set max for test only
+    uint32 _trustPriceAge = type(uint32).max; // set max for test only
     oracleMiddleWare.setAssetPriceConfig(wethAssetId, _confidenceThresholdE6, _trustPriceAge);
     oracleMiddleWare.setAssetPriceConfig(wbtcAssetId, _confidenceThresholdE6, _trustPriceAge);
     oracleMiddleWare.setAssetPriceConfig(daiAssetId, _confidenceThresholdE6, _trustPriceAge);
